refactor(dependency): sink get_merged_pr_for_issue to clients layer
Move PR status checker from services/pr/ to clients/ to eliminate
the shared->pr boundary violation and remove the callback injection
pattern from DependencyResolutionService.

- Create vibe3.clients.pr_status_checker with zero services imports
- Inline get_git_common_dir as GitClient().get_git_common_dir()
- Remove DependencyResolutionService.check_merged_pr_fn callback
- Simplify callers by removing 2 lines each of callback boilerplate
- Delete old services/pr/status_checker.py
