+++ src/vibe3/services/__init__.py
@@ -226,7 +226,6 @@ __all__ = [
     "format_agent_actor",
     "format_issue_runtime_line",
     "format_issue_summary_line",
-    "has_merged_pr_for_issue",
     "get_flow_state",
     "get_handoff_state_label",
     "get_manager_usernames",
@@ -330,7 +329,6 @@ _SYMBOL_MODULES = {
     "fetch_task_status_data": "vibe3.services.task.status",
     "filter_critical_files": "vibe3.services.pr.analysis",
     "format_agent_actor": "vibe3.services.shared.actors",
-    "has_merged_pr_for_issue": "vibe3.services.pr.status_checker",
     "format_issue_runtime_line": "vibe3.services.orchestra.status",
     "format_issue_summary_line": "vibe3.services.orchestra.status",
     "get_flow_state": "vibe3.services.flow.classifier",
