+++ src/vibe3/services/check/service.py
@@ -390,7 +390,6 @@ class CheckService(CheckRemote):
             # ---- read-only dependency check ----
             if task_issue and flow_status not in self.INACTIVE_FLOW_STATUSES:
                 from vibe3.services.orchestra.coordination import CoordinationResolver
-                from vibe3.services.pr import get_merged_pr_for_issue
                 from vibe3.services.shared import DependencyResolutionService
 
                 resolver = CoordinationResolver(store=self.store)
@@ -403,7 +402,6 @@ class CheckService(CheckRemote):
                             dep,
                             github_client=self.github_client,
                             repo=None,  # current CheckService doesn't carry repo
-                            check_merged_pr_fn=get_merged_pr_for_issue,
                         )
                         if not resolution.resolved:
                             unresolved.append(dep)
