+++ src/vibe3/services/task/resume.py
@@ -325,7 +325,7 @@ class TaskResumeCandidates:
             reason is a human-readable message if can_resume is False, None otherwise.
         """
         # ✅ Use authoritative truth: check if issue has merged PR
-        from vibe3.services.pr import has_merged_pr_for_issue
+        from vibe3.clients import has_merged_pr_for_issue
 
         if has_merged_pr_for_issue(issue_number, repo):
             # Issue has merged PR, cannot be resumed
@@ -374,14 +374,12 @@ class TaskResumeCandidates:
         if not isinstance(blocked_by_issue, int) or blocked_by_issue <= 0:
             return None
 
-        from vibe3.services.pr import get_merged_pr_for_issue
         from vibe3.services.shared import DependencyResolutionService
 
         resolution = DependencyResolutionService.is_dependency_resolved(
             blocked_by_issue,
             github_client=self._github_client,
             repo=repo,
-            check_merged_pr_fn=get_merged_pr_for_issue,
         )
         if not resolution.resolved:
             return (
