+++ tests/vibe3/services/test_check_pr_status.py
@@ -230,8 +230,8 @@ class TestMergedPRCacheIntegration:
 
     def test_cache_hit_skips_api_call(self, tmp_path: Path) -> None:
         """When cache hits, get_merged_pr_for_issue should skip API call."""
+        from vibe3.clients import get_merged_pr_for_issue
         from vibe3.clients.merged_pr_cache import MergedPRCache
-        from vibe3.services.pr.status_checker import get_merged_pr_for_issue
 
         cache = MergedPRCache(tmp_path)
         cache._save_cache(
@@ -267,7 +267,7 @@ class TestMergedPRCacheIntegration:
 
     def test_cache_miss_triggers_sync(self, tmp_path: Path) -> None:
         """When cache misses, get_merged_pr_for_issue should sync and return result."""
-        from vibe3.services.pr.status_checker import get_merged_pr_for_issue
+        from vibe3.clients import get_merged_pr_for_issue
 
         # Mock get_git_common_dir to return tmp_path
         with patch(
@@ -303,7 +303,7 @@ class TestMergedPRCacheIntegration:
 
     def test_cache_miss_records_sync_api_error(self, tmp_path: Path) -> None:
         """Cache sync API failures should be recorded, not silently swallowed."""
-        from vibe3.services.pr.status_checker import get_merged_pr_for_issue
+        from vibe3.clients import get_merged_pr_for_issue
 
         with patch(
             "vibe3.services.pr.status_checker.get_git_common_dir"
